raspberry_pi/requirements.txt
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
flask==3.1.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==3.0.2
pycparser==2.22
pycryptodome==3.21.0
requests==2.32.3
urllib3==2.2.3
werkzeug==3.1.3
zipp==3.21.0
